requirements.txt
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
requests>=2.31.0
jupyter>=1.0.0
python-dotenv>=1.0.0
ipykernel>=6.29.0